refactor: add feature-shell about page(list/detail)
